Add bis_pipeline stuff to pydrip
